+++ application/src/types/app.ts
@@ -77,6 +77,8 @@ export enum SupportedPaymentMethods {
   in3 = 'in3',
   alma = 'alma',
   payconiq = 'payconiq',
+  riverty = 'riverty',
+  swish = 'swish',
 }
 
 export const PaymentsWithResitrctedSurcharge = ['klarna', 'in3', 'alma'];