+++ job/src/utils/__mocks__/config.utils.ts
@@ -1,7 +1,0 @@
-export const readConfiguration = jest.fn(() => ({
-  clientId: 'mockedClientId',
-  clientSecret: 'mockedClientSecret',
-  projectKey: 'mockedProjectKey',
-  scope: 'mockedScope',
-  region: 'mockedRegion',
-}));