+++ processor/src/routes/processor.route.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import { post } from '../controllers/processor.controller';
 import { install, healthCheck, uninstall, mollieStatus } from '../controllers/connector.controller';
 import { readConfiguration } from '../utils/config.utils';
-import { oauthMiddleware } from '../middleware/oauth.middleware';
+import { basicAuthMiddleware } from '../middleware/basicAuth.middleware';
 
 const serviceRouter = Router();
 const AUTH_MODE = readConfiguration().commerceTools.authMode === '1';
@@ -12,9 +12,9 @@ serviceRouter.get('/health-check', healthCheck);
 serviceRouter.get('/mollie/status', mollieStatus);
 
 if (AUTH_MODE) {
-  serviceRouter.post('/', oauthMiddleware, post);
-  serviceRouter.post('/install', oauthMiddleware, install);
-  serviceRouter.post('/uninstall', oauthMiddleware, uninstall);
+  serviceRouter.post('/', basicAuthMiddleware, post);
+  serviceRouter.post('/install', basicAuthMiddleware, install);
+  serviceRouter.post('/uninstall', basicAuthMiddleware, uninstall);
 } else {
   serviceRouter.post('/', post);
   serviceRouter.post('/install', install);